npc code restructuring
